% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaryIPOT.R
\name{BaryIPOT}
\alias{BaryIPOT}
\title{Wasserstein Barycenter with IPOT Algorithm}
\usage{
BaryIPOT(
  list.data,
  list.weights = NULL,
  p = 1,
  L = 1,
  beta = 0.1,
  lambda = rep(1/length(list.data), length(list.data))
)
}
\description{
Wasserstein Barycenter with IPOT Algorithm
}
\examples{
## let's try a simple example
mydata = list()
for (i in 1:2){
  ni = round(stats::runif(1, min=10, max=20))
  mydata[[i]] = matrix(rnorm(ni*2),ncol=2)
}
hey = BaryIPOT(mydata)


}
