% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_digit4.R
\docType{data}
\name{digit4}
\alias{digit4}
\title{MNIST Images of Digit 4}
\format{
a length-\eqn{2000} named list \code{"digit4"} of \eqn{(28\times 28)} matrices.
}
\usage{
data(digit4)
}
\description{
\code{digit4} contains 2000 images from the famous MNIST dataset of digit 3. 
Each element of the list is an image represented as an \eqn{(28\times 28)} 
matrix that sums to 1. This normalization is conventional and it does not 
hurt its visualization via a basic `image()` function.
}
\examples{
## LOAD THE DATA
data(digit4)

## SHOW A FEW
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,4), pty="s")
for (i in 1:8){
  image(digit4[[i]])
}
par(opar)

}
\concept{data}
\keyword{datasets}
