% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gromov_gw.R
\name{gw}
\alias{gw}
\alias{gwD}
\title{Gromov-Wasserstein Distance}
\usage{
gw(X, Y, wx = NULL, wy = NULL, loss = c("square", "KL"), armijo = FALSE)

gwD(
  distX,
  distY,
  wx = NULL,
  wy = NULL,
  loss = c("square", "KL"),
  armijo = FALSE
)
}
\value{
a named list containing\describe{
\item{distance}{Gromov-Wasserstein distance value.}
\item{plan}{an \eqn{(M\times N)} nonnegative matrix for the optimal transport plan.}
}
}
\description{
Gromov-Wasserstein Distance
}
\examples{
#-------------------------------------------------------------------
#  Gromow-Wasserstein Distance between Two Displaced Bivariate Normal
#
# * class 1 : samples from Gaussian with mean=(-5, -5)
# * class 2 : samples from Gaussian with mean=(+5, +5)
#-------------------------------------------------------------------
if (reticulate::py_module_available("ot")){
  ## SMALL EXAMPLE
  m = sample(50:100, 1)
  n = sample(50:100, 1)
  X = matrix(rnorm(m*2, mean=-1),ncol=2) # m obs. for X
  Y = matrix(rnorm(n*2, mean=+1),ncol=2) # n obs. for Y

  xyloc = rbind(X, Y)
  xycol = c(rep(2,m),rep(4,n))

  ## COMPUTE WITH DIFFERENT COST
  out.sq = gw(X, Y, loss="Square")
  out.kl = gw(X, Y, loss="kl")

  ## VISUALIZE : SHOW THE PLAN AND DISTANCE
  pm.sq = paste0("Square loss; dist=",round(out.sq$distance,4))
  pm.kl = paste0("KL loss; dist=",round(out.kl$distance,4))

  opar <- par(no.readonly=TRUE)
  par(mfrow=c(1,3))
  plot(xyloc, col=xycol, pch=19, main="data")
  image(out.sq$plan, axes=FALSE, main=pm.sq)
  image(out.kl$plan, axes=FALSE, main=pm.kl)
  par(opar)
} else {
  print("No POT module detected. Use 'install_pot()' to run this example.")
}

}
\references{
\insertRef{memoli_gromov_2011}{T4transport}
}
\concept{gromov}
