% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gw_sinkhorn.R
\name{gwdist16P}
\alias{gwdist16P}
\title{Gromov-Wasserstein Distance via Entropic Regularization by}
\usage{
gwdist16P(C1, C2, p1 = NULL, p2 = NULL, cost = c("KL", "quadratic"), ...)
}
\description{
Gromov-Wasserstein Distance via Entropic Regularization by
}
\concept{dist_gw}
