% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pot.R
\name{install_pot}
\alias{install_pot}
\title{Install Python Optimal Transport Library}
\usage{
install_pot(method = "auto", conda = "auto")
}
\description{
Install Python Optimal Transport Library
}
\examples{
\dontrun{
T4transport::install_pot()
}

}
