% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPOT.R
\name{IPOT}
\alias{IPOT}
\title{IPOT}
\usage{
IPOT(
  dxy,
  p = 1,
  wx = NULL,
  wy = NULL,
  beta = 0.1,
  L = 1,
  maxiter = 496,
  abstol = 1e-06
)
}
\description{
IPOT
}
\examples{
## create two small datasets from bivariate normal
m = 20
n = 10
X = matrix(rnorm(m*2),ncol=2) # m obs. for X
Y = matrix(rnorm(n*2),ncol=2) # n obs. for Y

## compute cross-distance between X and Y
dXY = array(0,c(m,n))
for (i in 1:m){
  vx = as.vector(X[i,])
  for (j in 1:n){
    vy  = as.vector(Y[j,])
    dXY[i,j] = sqrt(sum((vx-vy)^2))
  }
}

## use several betas and compare with Wasserstein distance
outLP = Wasserstein(dXY, method="lp")
outP1 = IPOT(dXY, beta=0.01)
outP2 = IPOT(dXY, beta=0.05)
outP3 = IPOT(dXY, beta=0.1)

## visualize
#  show the transport plans and computed distance
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,2))
image(outLP$plan,main=paste0("LP:",round(outLP$distance,3)))
image(outP1$plan,main=paste0("beta=0.01:",round(outP1$distance,3)))
image(outLP$plan,main=paste0("beta=0.05:",round(outP2$distance,3)))
image(outLP$plan,main=paste0("beta=0.10:",round(outP3$distance,3)))
par(opar)

}
