% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaryIPOT.R
\name{IPOTWB.R}
\alias{IPOTWB.R}
\title{R version of IPOT-WB}
\usage{
IPOTWB.R(C, probs, beta = 0.1)
}
\description{
R version of IPOT-WB
}
\examples{
X = matrix(rnorm(100*2),ncol=2)
C = as.matrix(stats::dist(X))^2
probs = list()
for (k in 1:20){
  tmprobs = stats::runif(100)
  probs[[k]] = tmprobs/sum(tmprobs)
}

}
