% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_image22Y.R
\name{medimage22Y}
\alias{medimage22Y}
\title{Wasserstein Median of Images by You et al. (2022)}
\usage{
medimage22Y(images, weights = NULL, lambda = NULL, ...)
}
\description{
Wasserstein Median of Images by You et al. (2022)
}
\examples{
#----------------------------------------------------------------------
#                       MNIST Data with Digit 3
#
# EXAMPLE 1 : Very Small  Example for CRAN; just showing how to use it!
# EXAMPLE 2 : Medium-size Example for Evolution of Output
#----------------------------------------------------------------------
# EXAMPLE 1
data(digit3)
datsmall = digit3[1:3]
outsmall = medimage22Y(datsmall, print.progress=TRUE)


}
\concept{median_image}
