% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_bary_pw.R
\name{pwbary}
\alias{pwbary}
\title{Procrustes-Wasserstein Barycenter}
\usage{
pwbary(atoms, marginals = NULL, weights = NULL, num_support = 100, ...)
}
\description{
Procrustes-Wasserstein Barycenter
}
\concept{free_centroid}
