% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_median_GD.R
\name{rmedGD}
\alias{rmedGD}
\title{Free-Support Wasserstein Median by Riemannian gradient descent}
\usage{
rmedGD(X)
}
\description{
Free-Support Wasserstein Median by Riemannian gradient descent
}
\concept{free_median}
