% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_median_PF.R
\name{rmedPF}
\alias{rmedPF}
\title{Free-Support Median by Particle-Flow Algorithm}
\usage{
rmedPF(atoms, marginals = NULL, weights = NULL, num_support = 100, ...)
}
\description{
Free-Support Median by Particle-Flow Algorithm
}
\concept{free_centroid}
